+++ Controller/Repository.cs
@@ -4,10 +4,10 @@ using System.Linq;
 
 namespace BlogMvcCore.Storage
 {
-    public class Repository : IAuthenticationAction, IUserAction, IPostAction, ICommentAction
+    public class Repository : IAuthenticationAction, IUserAction, ICommentAction
     {
-        private readonly UserDbContext context;
-        public Repository(UserDbContext context) => this.context = context;
+        private readonly DbContext context;
+        public Repository(DbContext context) => this.context = context;
 
         public DomainModel.User FindUser(string login)
         {
@@ -16,29 +16,7 @@ namespace BlogMvcCore.Storage
             return new(user.FirstName, user.SecondName, user.Login, user.Password);
         }
 
-        public void AddPost(DomainModel.Post post)
-        {
-            User entityAuthor = context.BlogUsers.Where(u => u.Login == post.Author.Login).
-                                                  First();
-            context.Attach(entityAuthor);
-            context.Posts.Add(new Post()
-            {
-                ID = post.ID,
-                Author = entityAuthor,
-                Title = post.Title,
-                Text = post.Text,
-                Date = post.Date
-            });
-            context.SaveChanges();
-        }
 
-        public void DeletePost(long postID)
-        {
-            var deletingPost = context.Posts.FirstOrDefault(x => x.ID == postID);
-            if (deletingPost != null)
-                context.Entry(deletingPost).State = Microsoft.EntityFrameworkCore.EntityState.Deleted;
-            context.SaveChanges();
-        }
 
         public bool LoginUser(string login, string password)
         {
@@ -56,40 +34,7 @@ namespace BlogMvcCore.Storage
 
         public void Dispose() => context.Dispose();
 
-        public List<DomainModel.Post> GetUserPost(DomainModel.User user)
-        {
-            var entityPostsList = context.Posts.Where(p => p.Author.Login == user.Login)
-                                               .ToList();
 
-            List<DomainModel.Post> postsDomain = new();
-            foreach (var item in entityPostsList)
-            {
-                DomainModel.Post postDomain = new()
-                {
-                    ID = item.ID,
-                    Author = user,
-                    Title = item.Title,
-                    Text = item.Text,
-                    Date = item.Date
-                };
-                postsDomain.Add(postDomain);
-            }
-
-            return postsDomain.OrderByDescending(p => p.Date).ToList();
-        }
-
-        public DomainModel.Post FindPost(long postID)
-        {
-            var postStorage = context.Posts.Find(postID);
-
-            return new DomainModel.Post()
-            {
-                ID = postStorage.ID,
-                Title = postStorage.Title,
-                Text = postStorage.Text,
-                Date = postStorage.Date
-            }; ;
-        }
 
         public void AddComment(DomainModel.Comment comment)
         {
@@ -123,31 +68,6 @@ namespace BlogMvcCore.Storage
             context.SaveChanges();
         }
 
-        public List<DomainModel.Comment> GetPostComment(DomainModel.Post post)
-        {
-            var entityComments = context.Comments.Where(c => c.Post.ID == post.ID).ToList();
-
-            var commentList = entityComments.Select(c => new DomainModel.Comment
-            {
-                Post = new DomainModel.Post
-                {
-                    ID = post.ID,
-                    Author = post.Author,
-                    Title = post.Title,
-                    Text = post.Text,
-                    Date = post.Date,
-                    Comments = post.Comments
-                },
-                ID = c.ID,
-                Parent = c.Parent,
-                Author = c.Author,
-                Text = c.Text,
-                Date = c.Date
-            }).OrderByDescending(c => c.Date)
-              .ToList();
-
-            return commentList;
-        }
 
         public List<DomainModel.User> GetUsersList()
         {
@@ -156,24 +76,5 @@ namespace BlogMvcCore.Storage
             return userDomainList;
         }
 
-        public List<DomainModel.Post> GetPostList()
-        {
-            var entityPostsList = context.Posts.ToList().OrderByDescending(p => p.Date);
-
-            List<DomainModel.Post> postList = new();
-            foreach (var item in entityPostsList)
-            {
-                DomainModel.Post postDomain = new()
-                {
-                    ID = item.ID,
-                    Title = item.Title,
-                    Text = item.Text,
-                    Date = item.Date
-                };
-                postList.Add(postDomain);
-            }
-
-            return postList;
-        }
     }
 }
